git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@952 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/MEAM/pair_meam.cpp
@@ -890,9 +890,9 @@ void PairMEAM::unpack_reverse_comm(int n, int *list, double *buf)
    memory usage of local atom-based arrays 
 ------------------------------------------------------------------------- */
 
-int PairMEAM::memory_usage()
+double PairMEAM::memory_usage()
 {
-  int bytes = 7 * nmax * sizeof(double);
+  double bytes = 7 * nmax * sizeof(double);
   bytes += (3 + 6 + 10 + 3 + 3) * nmax * sizeof(double);
   bytes += 3*3 * nmax * sizeof(double);
   bytes += 2 * maxneigh * sizeof(double);
