+++ src/USER-ACKLAND/compute_ackland_atom.cpp
@@ -378,8 +378,8 @@ void ComputeAcklandAtom::select2(int k, int n, double *arr, int *iarr)
    memory usage of local atom-based array
 ------------------------------------------------------------------------- */
 
-int ComputeAcklandAtom::memory_usage()
+double ComputeAcklandAtom::memory_usage()
 {
-  int bytes = nmax * sizeof(double);
+  double bytes = nmax * sizeof(double);
   return bytes;
 }
